refactor : cctv 관련 기능 도메인 변경
