+++ BackEnd/src/main/java/com/example/ahpuh/cctv/dto/PostCctvPosListReq.java
@@ -1,14 +1,16 @@
-package com.example.ahpuh.admin.dto;
+package com.example.ahpuh.cctv.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PostCctvFileRes {
-    private String cctvImg;
+public class PostCctvPosListReq {
+    private List<PostCctvPosReq> posList;
 }
