+++ BackEnd/src/main/java/com/example/ahpuh/cctv/dto/PostLineReq.java
@@ -1,16 +1,14 @@
-package com.example.ahpuh.admin.dto;
+package com.example.ahpuh.cctv.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PostCctvPosListReq {
-    private List<PostCctvPosReq> posList;
+public class PostLineReq {
+    private Long lineNum;
 }
